requirements.txt
# Project Dependencies
python-telegram-bot[ext]
sqlalchemy
psycopg2-binary
apscheduler
python-dotenv
